fixed font vfs generation script on macos
On macOS the shell script does not produce usable output. The `-n` is present in the resulting js file which causes the newlines to be present as well which then causes the js file to be invalid. The "oneliner" echo version with variables does work in macOS, however. (I changed the `-b` back to `-w` here)

If this also works on other OS, it might be worth to change in the example code.